refactor: util -> common, spread out modules
