+++ ui/cli.py
@@ -1,4 +1,4 @@
-from util.app import App
+from common.app import App
 import time
